Fallback to osu!direct for cheesegull conns
